# [dipu] perf: use nodispatch::empty and nodispatch::empty_like to eliminate the dispatch of empty