--- a/dipu/torch_dipu/csrc_dipu/aten/ops/NodispatchUtils.hpp
+++ b/dipu/torch_dipu/csrc_dipu/aten/ops/NodispatchUtils.hpp
@@ -23,13 +23,23 @@ namespace nodispatch {
 inline at::Tensor empty(
     at::IntArrayRef size, at::TensorOptions options = {},
     c10::optional<at::MemoryFormat> memory_format = c10::nullopt) {
-  return dipu_aten::empty(
-      size, c10::optTypeMetaToScalarType(options.dtype_opt()),
-      options.layout_opt(), options.device_opt(), options.pinned_memory_opt(),
-      c10::impl::check_tensor_options_and_extract_memory_format(options,
-                                                                memory_format));
+  if (c10::device_or_default(options.device_opt()).type() ==
+      c10::DeviceType::CPU) {
+    return dipu_aten::empty_cpu(
+        size, c10::optTypeMetaToScalarType(options.dtype_opt()),
+        options.layout_opt(), options.device_opt(), options.pinned_memory_opt(),
+        c10::impl::check_tensor_options_and_extract_memory_format(
+            options, memory_format));
+  } else {
+    return dipu_aten::empty(
+        size, c10::optTypeMetaToScalarType(options.dtype_opt()),
+        options.layout_opt(), options.device_opt(), options.pinned_memory_opt(),
+        c10::impl::check_tensor_options_and_extract_memory_format(
+            options, memory_format));
+  }
 }
 
+// The code that calls this overloaded function is all for allocating CPU memory
 inline at::Tensor empty_cpu(
     at::IntArrayRef size, at::ScalarType dtype,
     c10::optional<at::Device> device_opt = c10::nullopt,
@@ -40,8 +50,26 @@ inline at::Tensor empty_cpu(
 }
 
 // an simplified version of `at::empty_like` but without dispatch
-inline at::Tensor empty_like(const at::Tensor& self) {
-  return empty(self.sizes(), self.options());
+inline at::Tensor empty_like(
+    const at::Tensor& self, at::TensorOptions options = {},
+    c10::optional<at::MemoryFormat> memory_format = c10::nullopt) {
+  if (!options.has_device()) {
+    return empty(self.sizes(), self.options(), memory_format);
+  } else {
+    return empty(self.sizes(), options, memory_format);
+  }
+}
+
+// In the current code, this interface has only been called once, including
+// dtype, layout, device, and pin_memoriy are all c10::nullopt, so just ignore
+// these parameters
+inline at::Tensor empty_like(const at::Tensor& self,
+                             c10::optional<at::ScalarType> dtype,
+                             c10::optional<at::Layout> layout,
+                             c10::optional<at::Device> device,
+                             c10::optional<bool> pin_memory,
+                             c10::optional<at::MemoryFormat> memory_format) {
+  return empty(self.sizes(), self.options(), memory_format);
 }
 
 }  // namespace nodispatch