--- a/dipu/torch_dipu/csrc_dipu/aten/ops/NodispatchUtils.hpp
+++ b/dipu/torch_dipu/csrc_dipu/aten/ops/NodispatchUtils.hpp
@@ -49,27 +49,59 @@ inline at::Tensor empty_cpu(
                               c10::get_contiguous_memory_format());
 }
 
+inline at::Tensor empty_like(
+    const at::Tensor& self, c10::optional<at::ScalarType> dtype,
+    c10::optional<at::Layout> layout, c10::optional<at::Device> device,
+    c10::optional<bool> pin_memory,
+    c10::optional<at::MemoryFormat> optional_memory_format) {
+  at::TensorOptions options_ = at::TensorOptions()
+                                   .dtype(dtype)
+                                   .layout(layout)
+                                   .device(device)
+                                   .pinned_memory(pin_memory);
+  at::TensorOptions options =
+      self.options().merge_in(options_).merge_memory_format(
+          optional_memory_format);
+
+  TORCH_CHECK(!(options.layout() != c10::kStrided &&
+                optional_memory_format.has_value()),
+              "memory format option is only supported by strided tensors");
+
+  auto memory_format =
+      options.memory_format_opt().value_or(at::MemoryFormat::Preserve);
+
+  at::Tensor result;
+
+  // TODO(liuweiyu): need to implement nodispatch::empty_stride
+  if (memory_format == at::MemoryFormat::Preserve) {
+    result =
+        empty(self.sizes(), options.memory_format(self.suggest_memory_format()),
+              c10::nullopt);
+  } else {
+    result =
+        empty(self.sizes(), options.memory_format(memory_format), c10::nullopt);
+  }
+
+  if (self.opt_names()) {
+    at::namedinference::propagate_names(result, self.names());
+  }
+
+  // never propagate Conjugate, Negative, and ZeroTensor dispatch key
+  result._set_conj(false);
+  result._set_neg(false);
+  result._set_zero(false);
+  return result;
+}
+
 // an simplified version of `at::empty_like` but without dispatch
 inline at::Tensor empty_like(
     const at::Tensor& self, at::TensorOptions options = {},
     c10::optional<at::MemoryFormat> memory_format = c10::nullopt) {
-  if (!options.has_device()) {
-    return empty(self.sizes(), self.options(), memory_format);
-  } else {
-    return empty(self.sizes(), options, memory_format);
-  }
-}
-
-// In the current code, this interface has only been called once, including
-// dtype, layout, device, and pin_memoriy are all c10::nullopt, so just ignore
-// these parameters
-inline at::Tensor empty_like(const at::Tensor& self,
-                             c10::optional<at::ScalarType> dtype,
-                             c10::optional<at::Layout> layout,
-                             c10::optional<at::Device> device,
-                             c10::optional<bool> pin_memory,
-                             c10::optional<at::MemoryFormat> memory_format) {
-  return empty(self.sizes(), self.options(), memory_format);
+  return nodispatch::empty_like(
+      self, c10::optTypeMetaToScalarType(options.dtype_opt()),
+      options.layout_opt(), options.device_opt(), options.pinned_memory_opt(),
+      c10::impl::check_tensor_options_and_extract_memory_format(options,
+                                                                memory_format));
 }
 
 }  // namespace nodispatch