| model            | pytorch | dipu before  | dipu after   | speedup |
|------------------|---------|--------------|--------------|---------|
| resnet18-bs1-fwd | 3.0ms   | 4.1ms(73.2%) | 3.8ms(78.9%) | 7.9%    |
| llama7b-eval-fwd | 70ms    | 94ms(74.5%)  | 94ms(74.5%)  | 0       |